+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 41.6.0
+
+**Misc**
+
+- Ajout de `frameborder="0"` dans les iframes miniplayer
+
 ## 41.5.4 (12/02/2026)
 
 **Misc**